+++ src/Contracts/PredefinedMonikers.cs
@@ -4,7 +4,7 @@
     {
         public const string Scroll = "Scroll";
         public const string Zoom = "Zoom";
-        //public const string IdeState = "State";
+        public const string Navigate = "Navigate";
         public const string Editor = "Editor";
     }
 }