+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [x] Output Window logging
 - [x] Add bookmark navigation
 - [x] Add debug navigation
+- [x] Reset zoom level on click
 - [ ] Add telemtry
 
 Features that have a checkmark are complete and available for